Feat(개인정보채널) : 피드 상세 조회

+++ backend/src/main/java/com/worq/worcation/domain/channel/domain/Feed.java
@@ -26,7 +26,7 @@ public class Feed {
     private String content;
 
     @Column(name = "heart")
-    private Long heart;
+    private int heart;
 
     @Column(name = "created_at")
     private Instant createdAt;
